Use CacheDatabaseService more frequently
