+++ functions/.nycrc.json
@@ -7,10 +7,10 @@
     "lib/tests",
     "**/*.test.js",
     "**/*.no-coverage.js",
-    "**/mock*.js"
+    "**/*.mock.js"
   ],
-  "branches": 40,
-  "lines": 40,
-  "functions": 40,
-  "statements": 40
+  "branches": 50,
+  "lines": 50,
+  "functions": 50,
+  "statements": 50
 }
