+++ functions/src/functions/message.ts
@@ -10,6 +10,7 @@ import { https } from 'firebase-functions'
 import { type CallableRequest, onCall } from 'firebase-functions/v2/https'
 import { FirestoreService } from '../services/database/firestoreService.js'
 import { DatabaseUserService } from '../services/user/databaseUserService.js'
+import { CacheDatabaseService } from '../services/database/cacheDatabaseService.js'
 
 export interface DismissMessageInput {
   messageId?: string
@@ -30,7 +31,9 @@ export const dismissMessageFunction = onCall(
       throw new https.HttpsError('invalid-argument', 'Message ID is required')
 
     try {
-      const service = new DatabaseUserService(new FirestoreService())
+      const service = new DatabaseUserService(
+        new CacheDatabaseService(new FirestoreService()),
+      )
       await service.dismissMessage(userId, messageId, didPerformAction ?? false)
     } catch (error) {
       console.error(error)
