+++ functions/src/services/user/userService.test.ts
@@ -8,35 +8,42 @@
 
 import { assert, expect } from 'chai'
 import admin from 'firebase-admin'
-import { FieldValue } from 'firebase-admin/firestore'
+import { FieldValue, Firestore } from 'firebase-admin/firestore'
 import { describe } from 'mocha'
-import { type Invitation } from '../models/invitation.js'
-import { type Admin, type Patient, type User } from '../models/user.js'
-import { FirestoreService } from '../services/database/firestoreService.js'
-import { DatabaseUserService } from '../services/user/databaseUserService.js'
-import { type UserService } from '../services/user/userService.js'
-import { type MockFirestore } from '../tests/mocks/firestore.js'
+import { type Invitation } from '../../models/invitation.js'
+import { type Admin, type Patient, type User } from '../../models/user.js'
+import { FirestoreService } from '../database/firestoreService.js'
+import { DatabaseUserService } from './databaseUserService.js'
+import { type UserService } from './userService.js'
+import { type MockFirestore } from '../../tests/mocks/firestore.js'
 import {
   cleanupMocks,
   setupMockAuth,
   setupMockFirestore,
-} from '../tests/setup.js'
+} from '../../tests/setup.js'
+import { UserMessage, UserMessageType } from '../../models/message.js'
+import { https } from 'firebase-functions'
+import { CacheDatabaseService } from '../database/cacheDatabaseService.js'
 
-describe('Functions: Invitation', () => {
+describe('UserService', () => {
+  let mockFirestore: MockFirestore
+  let userService: UserService
+  let firestore: Firestore
+
+  beforeEach(() => {
+    setupMockAuth()
+    mockFirestore = setupMockFirestore()
+    firestore = admin.firestore()
+    userService = new DatabaseUserService(
+      new CacheDatabaseService(new FirestoreService()),
+    )
+  })
+
+  afterEach(() => {
+    cleanupMocks()
+  })
+
   describe('enrollUser', () => {
-    let mockFirestore: MockFirestore
-    let userService: UserService
-
-    beforeEach(() => {
-      setupMockAuth()
-      mockFirestore = setupMockFirestore()
-      userService = new DatabaseUserService(new FirestoreService())
-    })
-
-    afterEach(() => {
-      cleanupMocks()
-    })
-
     it('enrolls an admin', async () => {
       const userId = 'mockAdminUserId'
       const invitationId = 'mockAdmin'
@@ -64,8 +71,6 @@ describe('Functions: Invitation', () => {
       const invitation = await userService.getInvitation(invitationId)
       if (!invitation) assert.fail('Invitation not found')
       await userService.enrollUser(invitation, userId)
-
-      const firestore = admin.firestore()
 
       const invitationSnapshot = await firestore
         .collection('invitations')
@@ -135,8 +140,6 @@ describe('Functions: Invitation', () => {
       const invitation = await userService.getInvitation(invitationId)
       if (!invitation) assert.fail('Invitation not found')
       await userService.enrollUser(invitation, userId)
-
-      const firestore = admin.firestore()
 
       const invitationSnapshot = await firestore
         .collection('invitations')
@@ -210,8 +213,6 @@ describe('Functions: Invitation', () => {
       if (!invitation) assert.fail('Invitation not found')
       await userService.enrollUser(invitation, userId)
 
-      const firestore = admin.firestore()
-
       const invitationSnapshot = await firestore
         .collection('invitations')
         .doc(invitationId)
@@ -281,8 +282,6 @@ describe('Functions: Invitation', () => {
       if (!invitation) assert.fail('Invitation not found')
       await userService.enrollUser(invitation, userId)
 
-      const firestore = admin.firestore()
-
       const invitationSnapshot = await firestore
         .collection('invitations')
         .doc(invitationId)
@@ -321,6 +320,48 @@ describe('Functions: Invitation', () => {
       expect(userData).to.exist
       expect(userData?.invitationCode).to.equal(invitationId)
       expect(userData?.dateOfEnrollment).to.equal(FieldValue.serverTimestamp())
+    })
+  })
+
+  describe('dismissMessage', () => {
+    it('should update the completionDate of messages', async () => {
+      const message: UserMessage = {
+        dueDate: new Date('2024-01-01'),
+        type: UserMessageType.medicationChange,
+        title: 'Medication Change',
+        description: 'You have a new medication!',
+        action: 'medications',
+        isDismissible: true,
+      }
+      await firestore.doc('users/mockUser/messages/0').set(message)
+      await userService.dismissMessage('mockUser', '0', true)
+      const updatedMessage = await firestore
+        .doc('users/mockUser/messages/0')
+        .get()
+      expect(updatedMessage.data()?.completionDate).to.not.be.undefined
+    })
+
+    it('should not update the completionDate of messages', async () => {
+      const message: UserMessage = {
+        dueDate: new Date('2024-01-01'),
+        type: UserMessageType.preAppointment,
+        title: 'Upcoming appointment',
+        description: 'You have an upcoming appointment!',
+        action: 'healthSummary',
+        isDismissible: false,
+      }
+      await firestore.doc('users/mockUser/messages/0').set(message)
+      try {
+        await userService.dismissMessage('mockUser', '0', true)
+        assert.fail('Message should not be dismissible.')
+      } catch (error) {
+        expect(error).to.deep.equal(
+          new https.HttpsError(
+            'invalid-argument',
+            'Message is not dismissible.',
+          ),
+        )
+      }
     })
   })
 })
